+++ src/card/cards/combat_constraints.rs
@@ -56,3 +56,5 @@ pub const WAYFARING_TEMPLE: CardDefinitionId = CardDefinitionId(1855);
 pub const SUNDERING_GROWTH: CardDefinitionId = CardDefinitionId(1856);
 pub const TROSTANI_SELESNYAS_VOICE: CardDefinitionId = CardDefinitionId(1857);
 pub const VITU_GHAZI_GUILDMAGE: CardDefinitionId = CardDefinitionId(1858);
+pub const MASTER_OF_THE_PEARL_TRIDENT: CardDefinitionId = CardDefinitionId(1859);
+pub const SLEEP: CardDefinitionId = CardDefinitionId(1860);